refactor(websocket) : ChatMessageDto location Change
